refactor: Move models nodes to individual files
Implements ML-597
